Code First flight
The only change made posterior to the first flight tests were the fact that the ESC runs now on OsPriorityRealtime

+++ 1_Code/3_STM32/0_workspace/X1_Thorn_v3/Core/Src/User/Peripheral/Servo/Servo.c
@@ -38,7 +38,7 @@ int Servo_Init(servo_t *servo, TIM_HandleTypeDef *htim, uint32_t channel, uint16
 }
 
 
-void Servo_Set_Position(const servo_t *servo, uint16_t position)
+void Servo_Set_Position(const servo_t *servo, int32_t position)
 {
 	uint16_t period = position + servo->setpoint;
