+++ 1_Code/3_STM32/0_workspace/X1_Thorn_v3/Core/Src/User/Peripheral/Servo/Servo.h
@@ -18,7 +18,7 @@ typedef struct {
 
 
 int Servo_Init(servo_t *servo, TIM_HandleTypeDef *htim, uint32_t channel, uint16_t setpoint);
-void Servo_Set_Position(const servo_t *servo, uint16_t position);
+void Servo_Set_Position(const servo_t *servo, int32_t position);
 
 
 #endif /* SRC_USER_PERIPHERAL_SERVO_SERVO_H_ */
